refactor: Consolidate decide code (#89)
* Consolidate decide code

There were two implementations of decide. That's not the end of the world, but I'm about to implement a new decide format (`/decide?v=4`) and I don't want to have to do it in two places.

So this adds a new file `decide.go` with a new interface `decider` and an implementation `decideClient` that both `Client` and `FeatureFlagsPoller` uses.

* Remove redundant client

* Use `io.ReadAll`

`ioutil.ReadAll` is deprecated.

* Pass the already created decider into the poller

* Move decide types into `decide.go`

* Remove unused method